[1/2] Moves files from helpers/controllers to resolvers directory
