+++ graphql/resolvers/challengesController.ts
@@ -3,10 +3,10 @@ import type {
   CreateChallengeMutationVariables,
   UpdateChallengeMutationVariables
 } from '../../graphql'
-import { lessons } from '../../graphql/resolvers/lessons'
+import { lessons } from './lessons'
 import prisma from '../../prisma'
-import { isAdminOrThrow } from '../isAdmin'
-import { validateLessonId } from '../validation/validateLessonId'
+import { isAdminOrThrow } from '../../helpers/isAdmin'
+import { validateLessonId } from '../../helpers/validation/validateLessonId'
 
 export const createChallenge = async (
   _parent: void,
